fix(ph): fix rllib unit test

+++ tests/utils/test_rllib.py
@@ -1,3 +1,4 @@
+import numpy as np
 import phantom as ph
 import pytest
 
@@ -55,7 +56,7 @@ def test_rllib_train_rollout(tmpdir):
     # Evaluate policy:
     results = ph.utils.rllib.evaluate_policy(
         directory=f"{tmpdir}/LATEST",
-        obs=ph.utils.ranges.LinspaceRange(0, 1, 2, name="r"),
+        obs=ph.utils.ranges.LinspaceRange(0, 1, 2, name="r", dtype=np.int),
         policy_id="mock_policy",
     )
     assert list(results) == [({"r": 0.0}, 0.0, 0), ({"r": 1.0}, 1.0, 0)]
